safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.6_Iteration_Statements/12.6.3_The_for_Statement/S12.6.3_A8.1_T3-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Unexpected identifier 'index'. Expected a ':' following the property name 'var'.
S12.6.3_A8.1_T3

PASS Expected parsing failure

TEST COMPLETE
